Rename demo adapter for testing use
